+++ packages/react-sprucebot/src/components/Button/Button.js
@@ -1,7 +1,36 @@
 import React, { Component } from 'react'
+import styled from 'styled-components'
 import PropTypes from 'prop-types'
 import Loader from '../Loader/Loader'
 
+const ButtonWrapper = styled.div`
+	display: flex;
+	width: 50%;
+	${props => props.left && `padding-right: 1.125em;`};
+	${props => props.right && `padding-left: 1.125em;`};
+`
+
+const StyledButton = styled.button`
+	${props =>
+		props.busy ||
+		(props.disabled &&
+			`
+			pointer-events: none;
+			cursor: not-allowed;
+		`)};
+`
+
+const StyledAnchor = styled.a`
+	${props =>
+		props.busy ||
+		(props.disabled &&
+			`
+		pointer-events: none;
+		cursor: not-allowed;
+	`)};
+`
+
+// TODO refactor into styled component
 export default class Button extends Component {
 	constructor(props) {
 		super(props)
@@ -9,6 +38,7 @@ export default class Button extends Component {
 			busy: !!props.busy
 		}
 	}
+
 	componentWillReceiveProps(nextProps) {
 		if (typeof nextProps.busy !== 'undefined') {
 			this.setState({
@@ -16,26 +46,34 @@ export default class Button extends Component {
 			})
 		}
 	}
+
 	onClick = e => {
-		if (this.props.onClick) {
-			this.props.onClick(e)
-		} else if (this.props.href) {
+		const { busy } = this.state
+		const { disabled, onClick, href, target, router } = this.props
+
+		if (busy || disabled) {
+			return
+		}
+
+		if (onClick) {
+			onClick(e)
+		} else if (href) {
 			e.preventDefault()
 			this.setState({ busy: true })
-			const url = this.props.href
+			const url = href
 			if (/^http/.test(url)) {
 				// If the href is a full domain name
-				if (this.props.target) {
-					window.open(url, this.props.target)
+				if (target) {
+					window.open(url, target)
 				} else {
 					window.open(url, '_self')
 				}
 			} else {
 				// Relative url
-				if (this.props.target) {
-					window.open(url, this.props.target)
-				} else if (this.props.router) {
-					this.props.router.push(url)
+				if (target) {
+					window.open(url, target)
+				} else if (router) {
+					router.push(url)
 				} else {
 					window.open(url, '_self')
 				}
@@ -48,6 +86,23 @@ export default class Button extends Component {
 			}, 10000)
 		}
 	}
+
+	renderView = () => {
+		const { busy } = this.state
+		const { hideLoader, loaderDark, loaderStyle, children } = this.props
+
+		if (busy && !hideLoader) {
+			return (
+				<Loader
+					dark={loaderDark ? true : false}
+					fullWidth={false}
+					loaderStyle={loaderStyle}
+				/>
+			)
+		}
+		return children
+	}
+
 	render() {
 		const {
 			tag,
@@ -67,6 +122,9 @@ export default class Button extends Component {
 			loaderStyle,
 			busy: propBusy,
 			hideLoader,
+			tertiary,
+			left,
+			right,
 			...props
 		} = this.props
 
@@ -102,23 +160,40 @@ export default class Button extends Component {
 		}
 
 		// if this button has a href or is a "remove" button, make it an anchor
-		const Tag = props.href || remove ? 'a' : tag
+		let Tag
+		if (props.href || remove) {
+			Tag = StyledAnchor
+		} else if (tag === 'button') {
+			Tag = StyledButton
+		} else {
+			Tag = tag
+		}
+
+		if (tertiary) {
+			return (
+				<ButtonWrapper left={left} right={right}>
+					<Tag
+						className={`${btnClass} ${className || ''}`}
+						onClick={this.onClick}
+						disabled={disabled}
+						busy={busy}
+						{...props}
+					>
+						{this.renderView()}
+					</Tag>
+				</ButtonWrapper>
+			)
+		}
 
 		return (
 			<Tag
 				className={`${btnClass} ${className || ''}`}
 				onClick={this.onClick}
+				disabled={disabled}
+				busy={busy}
 				{...props}
 			>
-				{busy && !hideLoader ? (
-					<Loader
-						dark={loaderDark ? true : false}
-						fullWidth={false}
-						loaderStyle={loaderStyle}
-					/>
-				) : (
-					children
-				)}
+				{this.renderView()}
 			</Tag>
 		)
 	}
@@ -134,7 +209,11 @@ Button.propTypes = {
 	href: PropTypes.string,
 	remove: PropTypes.bool,
 	toggle: PropTypes.bool,
-	hideLoader: PropTypes.bool
+	hideLoader: PropTypes.bool,
+	tertiary: PropTypes.bool,
+	left: PropTypes.bool,
+	right: PropTypes.bool,
+	type: PropTypes.string
 }
 
 Button.defaultProps = {