+++ packages/react-sprucebot/src/skillskit/store/actions/index.js
@@ -1,7 +1,9 @@
 import * as auth from './auth'
 import * as onboarding from './onboarding'
+import * as sharable from './sharable'
 
 module.exports = {
 	auth,
-	onboarding
+	onboarding,
+	sharable
 }