+++ packages/react-sprucebot/src/skillskit/store/reducers/index.js
@@ -1,6 +1,7 @@
 import auth from './auth'
 import config from './config'
 import onboarding from './onboarding'
+import sharable from './sharable'
 import { reducer as form } from 'redux-form'
 
-export default { auth, config, onboarding, form }
+export default { auth, config, onboarding, sharable, form }