+++ packages/sprucebot-skills-kit-server/middleware/auth.js
@@ -5,6 +5,8 @@ const config = require('config')
 module.exports = router => {
 	// jwt validation
 	const auth = async (id, ctx, next) => {
+		debug('AUTH CHECK', next, ctx, id)
+
 		if (!next) {
 			next = ctx
 			ctx = id
@@ -18,6 +20,31 @@ module.exports = router => {
 			try {
 				var decoded = jwt.verify(token, config.API_KEY.toString().toLowerCase())
 				ctx.auth = await ctx.sb.user(decoded.locationId, decoded.userId)
+
+				const realRole = ctx.auth.role
+
+				// Allow override of role if in dev mode
+				if (config.DEV_MODE) {
+					const devRole =
+						ctx.cookies.get('devRole') || ctx.request.headers['x-dev-role']
+
+					// Allow users to downgrade their role for testing purposes
+					switch (devRole) {
+						case 'teammate':
+							if (realRole === 'owner') {
+								ctx.auth.role = devRole
+							}
+
+						case 'guest':
+							if (realRole === 'owner' || realRole === 'teammate') {
+								ctx.auth.role = devRole
+							}
+
+						// By default just use the (real) assigned role
+						default:
+							break
+					}
+				}
 				ctx.auth.jwt = token
 				debug(`middleware/auth token valid`)
 			} catch (err) {