+++ packages/react-sprucebot/CHANGELOG.md
@@ -3,6 +3,73 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="6.18.0"></a>
+# [6.18.0](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/compare/v6.17.0...v6.18.0) (2018-07-11)
+
+
+### Bug Fixes
+
+* Calendar class typo ([a29340b](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/commit/a29340b))
+* Remove CL ([2d71738](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/commit/2d71738))
+
+
+### Features
+
+* import fix ([e0fcd11](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/commit/e0fcd11))
+
+
+
+
+
+<a name="6.17.0"></a>
+# [6.17.0](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/compare/v6.16.0...v6.17.0) (2018-07-11)
+
+
+### Features
+
+* logging test ([f02e53f](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/commit/f02e53f))
+* style test ([90f744f](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/commit/90f744f))
+
+
+
+
+
+<a name="6.15.1"></a>
+## [6.15.1](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/compare/v6.15.0...v6.15.1) (2018-07-11)
+
+
+### Bug Fixes
+
+* Only import babel-polyfill once ([5686a11](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/commit/5686a11))
+
+
+
+
+
+<a name="6.15.0"></a>
+# [6.15.0](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/compare/v6.14.0...v6.15.0) (2018-07-11)
+
+
+### Features
+
+* sharable components ([fcd4943](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/commit/fcd4943))
+
+
+
+
+
+<a name="6.14.0"></a>
+# [6.14.0](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/compare/v6.13.0...v6.14.0) (2018-07-09)
+
+
+### Features
+
+* Latest components and functionality ([8dd2623](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/commit/8dd2623))
+
+
+
+
+
 <a name="6.13.0"></a>
 # [6.13.0](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/compare/v6.12.2...v6.13.0) (2018-07-06)