+++ packages/sprucebot-skills-kit/CHANGELOG.md
@@ -3,6 +3,66 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="6.18.0"></a>
+# [6.18.0](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/compare/v6.17.0...v6.18.0) (2018-07-11)
+
+**Note:** Version bump only for package sprucebot-skills-kit
+
+
+
+
+
+<a name="6.17.0"></a>
+# [6.17.0](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/compare/v6.16.0...v6.17.0) (2018-07-11)
+
+**Note:** Version bump only for package sprucebot-skills-kit
+
+
+
+
+
+<a name="6.16.0"></a>
+# [6.16.0](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/compare/v6.15.1...v6.16.0) (2018-07-11)
+
+**Note:** Version bump only for package sprucebot-skills-kit
+
+
+
+
+
+<a name="6.15.1"></a>
+## [6.15.1](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/compare/v6.15.0...v6.15.1) (2018-07-11)
+
+**Note:** Version bump only for package sprucebot-skills-kit
+
+
+
+
+
+<a name="6.15.0"></a>
+# [6.15.0](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/compare/v6.14.0...v6.15.0) (2018-07-11)
+
+
+### Features
+
+* sharable components ([fcd4943](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/commit/fcd4943))
+
+
+
+
+
+<a name="6.14.0"></a>
+# [6.14.0](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/compare/v6.13.0...v6.14.0) (2018-07-09)
+
+
+### Features
+
+* Latest components and functionality ([8dd2623](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/commit/8dd2623))
+
+
+
+
+
 <a name="6.13.0"></a>
 # [6.13.0](https://github.com/sprucelabsai/workspace.sprucebot-skills-kit/compare/v6.12.2...v6.13.0) (2018-07-06)